requirements.txt
Django==3.2.12
djangorestframework==3.13.1
Markdown==3.3.6
Pillow==9.0.1


